Change quaternion order from wxyz to xyzw
I didn't know which order to choose and I had little reasons
for either or the other.
The only reason so far was glTF, which uses xyzw and I need to
swizzle during loading.
But now that I added ung_transform_set_* that take pointers
and I frequently pass &vec.x I thought it would be good if you
had to pass &quat.x as well instead of &quat.w (to avoid typo potential).
So I bit the bullet, because the engine is quite early still
and if I can ever change it, it's now.