Add very simple and basic texture caching
This should be guarded in devmode at some point (which does not
exist in ung yet). Also you have to create .ungcache for it to
wotk.

But in my example projects texture decoding is by far the slowest
part and is the major bottleneck for iteration time.
With texture caching I can get my load time from 1700ms (uncached)
to 350ms (cached)!